oh borrow checker

+++ src/selection.rs
@@ -1,12 +1,12 @@
 use macroquad::prelude::*;
 pub struct Selection {
-    max_index: i32,
+    max_index: usize,
     pub x: i32,
     pub y: i32,
 }
 
 impl Selection {
-    pub fn new(side_len_in_tiles: i32) -> Self {
+    pub fn new(side_len_in_tiles: usize) -> Self {
         Self {
             max_index: side_len_in_tiles - 1,
             x: 0,
@@ -21,7 +21,7 @@ impl Selection {
     }
 
     pub fn down(&mut self) {
-        if self.y < self.max_index {
+        if self.y < self.max_index as i32 {
             self.y += 1;
         }
     }
@@ -33,7 +33,7 @@ impl Selection {
     }
 
     pub fn right(&mut self) {
-        if self.x < self.max_index {
+        if self.x < self.max_index as i32 {
             self.x += 1;
         }
     }
